+++ airflow/datadog_checks/airflow/airflow.py
@@ -10,9 +10,7 @@ AIRFLOW_STATUS_OK = "OK"
 
 class AirflowCheck(AgentCheck):
     def __init__(self, name, init_config, instances):
-        super(AirflowCheck, self).__init__(
-            name, init_config, instances,
-        )
+        super(AirflowCheck, self).__init__(name, init_config, instances)
 
         self._url = self.instance.get('url', '')
         self._tags = self.instance.get('tags', [])