+++ ibm_mq/datadog_checks/ibm_mq/collectors/__init__.py
@@ -2,6 +2,7 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from .channel_metric_collector import ChannelMetricCollector
+from .metadata_collector import MetadataCollector
 from .queue_metric_collector import QueueMetricCollector
 
-__all__ = ['ChannelMetricCollector', 'QueueMetricCollector']
+__all__ = ['ChannelMetricCollector', 'MetadataCollector', 'QueueMetricCollector']