+++ snmp/datadog_checks/snmp/parsing/metrics.py
@@ -4,7 +4,8 @@
 """
 Helpers for parsing the `metrics` section of a config file.
 """
-from typing import Dict, List, NamedTuple, Sequence, TypedDict, Union, cast
+from logging import Logger
+from typing import Dict, List, NamedTuple, Optional, Sequence, TypedDict, Union, cast
 
 from datadog_checks.base import ConfigurationError
 
@@ -30,8 +31,8 @@ ParseMetricsResult = TypedDict(
 )
 
 
-def parse_metrics(metrics, resolver, bulk_threshold=0):
-    # type: (List[Metric], OIDResolver, int) -> ParseMetricsResult
+def parse_metrics(metrics, resolver, logger, bulk_threshold=0):
+    # type: (List[Metric], OIDResolver, Optional[Logger], int) -> ParseMetricsResult
     """
     Parse the `metrics` section of a config file, and return OIDs to fetch and metrics to submit.
     """
@@ -41,7 +42,7 @@ def parse_metrics(metrics, resolver, bulk_threshold=0):
     parsed_metrics = []  # type: List[ParsedMetric]
 
     for metric in metrics:
-        result = _parse_metric(metric)
+        result = _parse_metric(metric, logger)
 
         for oid in result.oids_to_fetch:
             oids.append(oid)
@@ -88,8 +89,8 @@ MetricParseResult = NamedTuple(
 )
 
 
-def _parse_metric(metric):
-    # type: (Metric) -> MetricParseResult
+def _parse_metric(metric, logger):
+    # type: (Metric, Optional[Logger]) -> MetricParseResult
     """
     Parse a single metric in the `metrics` section of a config file.
 
@@ -142,7 +143,7 @@ def _parse_metric(metric):
         if 'symbols' not in metric:
             raise ConfigurationError('When specifying a table, you must specify a list of symbols')
         metric = cast(TableMetric, metric)
-        return _parse_table_metric(metric)
+        return _parse_table_metric(metric, logger)
 
     raise ConfigurationError('When specifying a MIB, you must specify either a table or a symbol')
 
@@ -164,7 +165,11 @@ def _parse_oid_metric(metric):
     oid = OID(metric['OID'])
 
     parsed_symbol_metric = ParsedSymbolMetric(
-        name, tags=metric.get('metric_tags', []), forced_type=metric.get('forced_type'), enforce_scalar=False
+        name,
+        tags=metric.get('metric_tags', []),
+        forced_type=metric.get('forced_type'),
+        enforce_scalar=False,
+        options=metric.get('options', {}),
     )
 
     return MetricParseResult(
@@ -195,7 +200,10 @@ def _parse_symbol_metric(metric):
     parsed_symbol = _parse_symbol(mib, symbol)
 
     parsed_symbol_metric = ParsedSymbolMetric(
-        parsed_symbol.name, tags=metric.get('metric_tags', []), forced_type=metric.get('forced_type')
+        parsed_symbol.name,
+        tags=metric.get('metric_tags', []),
+        forced_type=metric.get('forced_type'),
+        options=metric.get('options', {}),
     )
 
     return MetricParseResult(
@@ -238,8 +246,8 @@ def _parse_symbol(mib, symbol):
     return ParsedSymbol(name=name, oid=oid, oids_to_resolve={name: oid})
 
 
-def _parse_table_metric(metric):
-    # type: (TableMetric) -> MetricParseResult
+def _parse_table_metric(metric, logger):
+    # type: (TableMetric, Optional[Logger]) -> MetricParseResult
     mib = metric['MIB']
 
     parsed_table = _parse_symbol(mib, metric['table'])
@@ -253,26 +261,35 @@ def _parse_table_metric(metric):
     index_mappings = []
     table_batches = {}  # type: TableBatches
 
-    for metric_tag in metric.get('metric_tags', []):
-        parsed_table_metric_tag = _parse_table_metric_tag(mib, parsed_table, metric_tag)
+    if metric.get('metric_tags'):
+        for metric_tag in metric['metric_tags']:
+            parsed_table_metric_tag = _parse_table_metric_tag(mib, parsed_table, metric_tag)
 
-        if isinstance(parsed_table_metric_tag, ParsedColumnMetricTag):
-            oids_to_resolve.update(parsed_table_metric_tag.oids_to_resolve)
-            column_tags.extend(parsed_table_metric_tag.column_tags)
-            table_batches = merge_table_batches(table_batches, parsed_table_metric_tag.table_batches)
+            if isinstance(parsed_table_metric_tag, ParsedColumnMetricTag):
+                oids_to_resolve.update(parsed_table_metric_tag.oids_to_resolve)
+                column_tags.extend(parsed_table_metric_tag.column_tags)
+                table_batches = merge_table_batches(table_batches, parsed_table_metric_tag.table_batches)
 
-        else:
-            index_tags.extend(parsed_table_metric_tag.index_tags)
+            else:
+                index_tags.extend(parsed_table_metric_tag.index_tags)
 
-            for index, mapping in parsed_table_metric_tag.index_mappings.items():
-                # Need to do manual resolution.
-                for symbol in metric['symbols']:
-                    index_mappings.append(IndexMapping(symbol['name'], index=index, mapping=mapping))
+                for index, mapping in parsed_table_metric_tag.index_mappings.items():
+                    # Need to do manual resolution.
+                    for symbol in metric['symbols']:
+                        index_mappings.append(IndexMapping(symbol['name'], index=index, mapping=mapping))
 
-                for tag in metric.get('metric_tags', []):
-                    if 'column' in tag:
-                        tag = cast(ColumnTableMetricTag, tag)
-                        index_mappings.append(IndexMapping(tag['column']['name'], index=index, mapping=mapping))
+                    for tag in metric.get('metric_tags', []):
+                        if 'column' in tag:
+                            tag = cast(ColumnTableMetricTag, tag)
+                            index_mappings.append(IndexMapping(tag['column']['name'], index=index, mapping=mapping))
+    elif logger:
+        logger.warning(
+            "%s table doesn't have a 'metric_tags' section, all its metrics will use the same tags. "
+            "If the table has multiple rows, only one row will be submitted. "
+            "Please add at least one discriminating metric tag (such as a row index) "
+            "to ensure metrics of all rows are submitted.",
+            str(metric['table']),
+        )
 
     # Then process symbols in the table.
 
@@ -286,7 +303,11 @@ def _parse_table_metric(metric):
         table_oids.append(parsed_symbol.oid)
 
         parsed_table_metric = ParsedTableMetric(
-            parsed_symbol.name, index_tags=index_tags, column_tags=column_tags, forced_type=metric.get('forced_type'),
+            parsed_symbol.name,
+            index_tags=index_tags,
+            column_tags=column_tags,
+            forced_type=metric.get('forced_type'),
+            options=metric.get('options', {}),
         )
         parsed_metrics.append(parsed_table_metric)
 
@@ -417,16 +438,7 @@ def _parse_table_metric_tag(mib, parsed_table, metric_tag):
 
 def _parse_column_metric_tag(mib, parsed_table, metric_tag):
     # type: (str, ParsedSymbol, ColumnTableMetricTag) -> ParsedColumnMetricTag
-    column = metric_tag['column']
-
-    if isinstance(column, dict) and 'table' in column:
-        # Prevent a common error due to bad indentation...
-        raise ConfigurationError(
-            'found unexpected "table" key in column {} '
-            '("table" should be set on the `metric_tag` - this is probably an indentation issue)'.format(column)
-        )
-
-    parsed_column = _parse_symbol(mib, column)
+    parsed_column = _parse_symbol(mib, metric_tag['column'])
 
     batches = {TableBatchKey(mib, table=parsed_table.name): TableBatch(parsed_table.oid, oids=[parsed_column.oid])}