+++ postgres/CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG - postgres
 
+## 5.0.2 / 2020-08-10
+
+* [Fixed] Update logs config service field to optional. See [#7209](https://github.com/DataDog/integrations-core/pull/7209).
+
+## 5.0.1 / 2020-07-16
+
+* [Fixed] Avoid aurora pg warnings. See [#7123](https://github.com/DataDog/integrations-core/pull/7123).
+
 ## 5.0.0 / 2020-06-29
 
 * [Added] Add config specs. See [#6547](https://github.com/DataDog/integrations-core/pull/6547).